ctutils: remove `(Partial)Eq` impls for `Choice` (#1315)
These were added in #1266 to simplify a migration from `subtle::Choice`
with a TODO to eventually remove them. They're used in tests, including
it seems, the ones for `ctutils`.

They're problematic because the goal of `Choice` is to be an opaque
boolean-alternative for use in constant-time code, but especially a
derived `Partial(Eq)` can peek inside them and bypass the encapsulation
they're trying to provide in a way that's easy to branch on.

Now that `crypto-bigint` has actually been migrated to `ctutils`, we can
followup on removing these as part of some final breaking changes.